readme pt. 3: new build
