+++ distroduce/configuration/launch_cluster.txt
